better-eng: separated settings
