+++ main_server.py
@@ -1,7 +1,24 @@
-import settings  # noqa
-from src.app import server
+import uvicorn
+from fastapi import FastAPI
 
-api_app = server.init_app()
+import settings_server  # noqa
+from src.core.server.handlers import exc_handler
+from src.core.server.routers import root_router
+
+
+def init_app():
+    app = FastAPI(title="Midjourney API")
+    
+    app.include_router(root_router, prefix=settings_server.API_PREFIX)
+    exc_handler(app)
+    
+    return app
+
+
+def run(host, port):
+    app = init_app()
+    uvicorn.run(app, port=port, host=host)
+
 
 if __name__ == '__main__':
-    server.run("0.0.0.0", 8062)
+    run("0.0.0.0", 8062)
