+++ src/core/server/handlers.py
@@ -1,0 +1,22 @@
+from fastapi import FastAPI
+from fastapi.exceptions import RequestValidationError
+from starlette import status
+from starlette.responses import JSONResponse
+
+from src.lib.exceptions import APPBaseException
+
+
+def exc_handler(app: FastAPI):
+    @app.exception_handler(RequestValidationError)
+    def validation_exception_handler(_, exc: RequestValidationError):
+        return JSONResponse(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            content=f"request params error: {exc.body}",
+        )
+    
+    @app.exception_handler(APPBaseException)
+    def validation_exception_handler(_, exc: APPBaseException):
+        return JSONResponse(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            content=exc.message,
+        )
