+++ src/ds/generic.py
@@ -1,0 +1,3 @@
+from typing import TypeVar
+
+T = TypeVar("T")
