+++ src/lib/callback.py
@@ -2,19 +2,22 @@ import json
 import time
 
 import aiohttp
-from loguru import logger
 
-import settings
+import settings_discord
+import src.lib.path
+from src.ds.discord import CallbackDict
 from src.lib.fetch import fetch
+from src.lib.log import logger
 
 
-async def callback(data):
-    if settings.DUMP_CALLBACK_DATA:
-        with open(settings.DATA_DIR / f"callback-{time.time()}.json", "w") as f:
+async def callback(data: CallbackDict):
+    if settings_discord.DUMP_CALLBACK_DATA:
+        with open(src.lib.path.DATA_DIR / f"callback-{time.time()}.json", "w") as f:
             json.dump(data, f, ensure_ascii=False, indent=2)
     
     logger.debug(f"callback data: {data}")
-    if not settings.CALLBACK_URL:
+    if not settings_discord.CALLBACK_URL:
+        logger.warning("没有配置 CALLBACK_URL，因此忽略回调数据")
         return
     
     headers = {"Content-Type": "application/json"}
@@ -22,4 +25,4 @@ async def callback(data):
         timeout=aiohttp.ClientTimeout(total=30),
         headers=headers
     ) as session:
-        await fetch(session, settings.CALLBACK_URL, json=data)
+        await fetch(session, settings_discord.CALLBACK_URL, json=data)
